requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
jinja2==3.1.6
pandas==2.3.2
openpyxl==3.1.5
PyMuPDF==1.26.4
openai==1.107.1
python-dotenv==1.1.1
rapidfuzz==3.14.0
